postblock3/Q6/requirements.txt
google-cloud-pubsub==2.21.5
protobuf>=4.25.0
